+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 * Unable to change 2SV method on Chrome when caching is enabled either at the Joomla or web server level.
 * Joomla 4 throws an exception when mail is disabled and you try to send an email
+* Unhandled exception page was incompatible with Joomla 4
 
 # 3.3.0